pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "django-nepkit"
version = "0.1.0"
description = "Django Nepali date, time, phone, and address fields with helpers."
readme = "README.md"
authors = [
{ name = "Sankalp Tharu", email = "sankalptharu50028@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "django>=4.2",
    "nepali>=1.1.3",
]
keywords = [
  "django",
  "django-fields",
  "django-forms",
  "nepal",
  "nepali",
  "nepali-date",
  "bikram-sambat",
  "nepali-calendar",
  "timezone",
  "asia-kathmandu",
  "phone-number",
  "address",
  "province",
  "district",
  "municipality"
]
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 6.0",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Operating System :: OS Independent",
    "Natural Language :: English"
]

[project.optional-dependencies]
drf = [
    "djangorestframework>=3.14",
]

[project.urls]
Homepage = "https://github.com/S4NKALP/django-nepkit"



[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
]
